+++ common/src/main/java/org/orecruncher/dsurround/lib/scripting/Script.java
@@ -3,7 +3,7 @@ package org.orecruncher.dsurround.lib.scripting;
 import com.google.common.base.Preconditions;
 import com.mojang.serialization.Codec;
 import org.jetbrains.annotations.Nullable;
-import org.orecruncher.dsurround.lib.scripting.engine.Expression;
+import org.orecruncher.dsurround.lib.scripting.engine.expression.Expression;
 
 import java.util.Optional;