+++ common/src/main/java/org/orecruncher/dsurround/runtime/variables/SeasonVariables.java
@@ -13,6 +13,9 @@ public final class SeasonVariables extends VariableSet {
     private boolean isSummer;
     private boolean isAutumn;
     private boolean isWinter;
+    private boolean isEarly;
+    private boolean isMiddle;
+    private boolean isLate;
 
     public SeasonVariables(ISeasonalInformation seasonalInformation) {
         super("season");
@@ -26,19 +29,29 @@ public final class SeasonVariables extends VariableSet {
             this.isSummer = this.seasonalInformation.isSummer();
             this.isAutumn = this.seasonalInformation.isAutumn();
             this.isWinter = this.seasonalInformation.isWinter();
+            this.isEarly = this.seasonalInformation.isEarly();
+            this.isMiddle = this.seasonalInformation.isMiddle();
+            this.isLate = this.seasonalInformation.isLate();
         } else {
             this.isSpring = false;
             this.isSummer = false;
             this.isAutumn = false;
             this.isWinter = false;
+            this.isEarly = false;
+            this.isMiddle = false;
+            this.isLate = false;
         }
     }
 
     @Override
     public void configure(IConfigureDefinition config) {
         config.defineFunction(id("isSpring"), l -> this.isSpring);
-        config.defineFunction(id("isSummer"), l -> this.isSummer);
+        config.defineFunction(id("isSummer"),l -> this.isSummer);
         config.defineFunction(id("isAutumn"), l -> this.isAutumn);
         config.defineFunction(id("isWinter"), l -> this.isWinter);
+
+        config.defineFunction(id("isEarly"), l -> this.isEarly);
+        config.defineFunction(id("isMiddle"), l -> this.isMiddle);
+        config.defineFunction(id("isLate"), l -> this.isLate);
     }
 }